Add files via upload
This algorithm maximally leverages the efficient algorithms available to compute the loop-
free version of the Reeb graph. It partitions the input into subsets whose Reeb graph do not
contain any loops and merges the loop-free graphs in linear time to construct the Reeb graph
of the input. The algorithm has a running time of O(t log t + l + n), which in the worst-case is
O(t log t +sn) or close to the lower bound Ω(t log t + n) as the size of critical level sets is
usually O(n) in practice and s is the number of saddles. Here t is the number of vertices or
critical points, n is the number of triangles, and l is the total size (number of edges) of all
critical level sets. In practice, the running times scale linearly with the number of triangles.
The algorithm is up to two orders of magnitude faster than existing generic algorithms that
support higher dimensional and non-manifold data. The performance is on par with
algorithms that are catered to restricted classes of input. The algorithm is also amenable to
handle large data that do not fit in memory. It supports 2D, 3D, and higher dimensional
meshes as input.